topologies: Update to work with QEMU > 5.2
QEMU commit
  b21aa7e01eb8 ("numa: drop support for '-numa node' (without memory specified)")
dropped support for automatically splitting memory between the nodes. It's
a bit of a PITA as we now need to do this manually, but that has to be done
if we want to run with newer QEMU versions.

This can be scriptified eventually, but just hardcode the memory split for now.